Add files via upload
In this activity, i am a member of an analytics team that provides insights to an investing firm. To help them decide which companies to invest in next, the firm wants insights into unicorn companies–companies that are valued at over one billion dollars. The data i will use for this task provides information on over 1,000 unicorn companies, including their industry, country, year founded, and select investors. i will use this information to gain insights into how and when companies reach this prestigious milestone and to make recommentations for next steps to the investing firm.